# sage-server/swagger-spec.json
{"openapi":"3.0.0","info":{"title":"Sage Recruitment","description":"Simple List API with Nest.js","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"ListItemEntity":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"order":{"type":"number"},"listId":{"type":"string"}},"required":["id","title","order","listId"]},"ListEntity":{"type":"object","properties":{"id":{"type":"string"},"items":{"type":"array","items":{"$ref":"#/components/schemas/ListItemEntity"}}},"required":["id","items"]},"AddItemDTO":{"type":"object","properties":{"listId":{"type":"string"},"title":{"type":"string"}},"required":["listId","title"]},"RemoveItemDTO":{"type":"object","properties":{"itemId":{"type":"string"}},"required":["itemId"]},"ClearListDTO":{"type":"object","properties":{"listId":{"type":"string"}},"required":["listId"]},"ReorderItem":{"type":"object","properties":{"id":{"type":"number"},"order":{"type":"number"}},"required":["id","order"]},"ReorderListDTO":{"type":"object","properties":{"listId":{"type":"number"},"items":{"type":"array","items":{"$ref":"#/components/schemas/ReorderItem"}}},"required":["listId","items"]}}},"paths":{"/list":{"get":{"operationId":"ListController_findAll","parameters":[],"responses":{"undefined":{"description":"All Lists.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ListEntity"}}}}}},"tags":["List"]},"post":{"operationId":"ListController_post","parameters":[],"responses":{"201":{"description":"Created List.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListEntity"}}}}},"tags":["List"]}},"/list/{id}":{"get":{"operationId":"ListController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Single list.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListEntity"}}}}},"tags":["List"]}},"/list/addItem":{"post":{"operationId":"ListController_addItem","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddItemDTO"}}}},"responses":{"201":{"description":"Created Item.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListItemEntity"}}}}},"tags":["List"]}},"/list/removeItem":{"delete":{"operationId":"ListController_removeItem","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RemoveItemDTO"}}}},"responses":{"undefined":{"description":"Boolean success","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["List"]}},"/list/clear":{"delete":{"operationId":"ListController_clear","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClearListDTO"}}}},"responses":{"200":{"description":"Boolean success","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["List"]}},"/list/reorder":{"post":{"operationId":"ListController_reorder","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReorderListDTO"}}}},"responses":{"200":{"description":"Reordered List","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListEntity"}}}}},"tags":["List"]}}}}